{{cookiecutter.project_slug}}/pyproject.toml
[project]
name = "{{cookiecutter.project_slug}}"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "sounddevice",
    "numpy",
    "pywebview",
    "soundfile",
]

[project.scripts]
dsp = "main:main"
